# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Oblivion API
  version: "1.0.0"
  description: |
    REST API for **Oblivion**, a face / genetic-syndrome recognition backend.

    The system is split into 7 independent microservices. They sit behind an
    nginx **API gateway** at `http://localhost:8080`, which routes by path
    prefix to the right service — so this whole API is one origin and
    "Try it out" works without switching ports. Endpoints are grouped by the
    **tag** (= service); the direct per-service ports are noted on each tag for
    when you want to bypass the gateway.

    ## Conventions
    - Every response is wrapped in an envelope:
      `{ "status": "ok" | "error", "data": <payload> | null, "error": <string> | null }`.
    - Authenticated endpoints expect `Authorization: Bearer <access_token>`,
      a RS256 JWT issued by the auth service.
    - Roles: `doctor`, `lab`, `admin`. Lab endpoints require a `lab` token.

    ## Typical flow
    1. `POST /auth/register` then `POST /auth/login` → access token.
    2. `POST /cases` → returns a presigned S3 `upload_url`.
    3. `PUT` the image bytes to `upload_url` (MinIO/S3, no auth).
    4. `POST /cases/{id}/confirm-upload` → triggers async analysis (Kafka).
    5. Poll `GET /cases/{id}` until `status = completed`; read `diagnoses`.
servers:
  - url: http://localhost:8080
    description: API gateway (nginx) — routes by path prefix to all services
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/User' } }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in and obtain tokens
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/AuthTokens' } }
        '401': { $ref: '#/components/responses/Error' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties: { refresh_token: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/AuthTokens' } }
        '401': { $ref: '#/components/responses/Error' }
  /auth/me:
    get:
      tags: [auth]
      summary: Get the current user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/User' } }
        '401': { $ref: '#/components/responses/Error' }
  /auth/logout:
    delete:
      tags: [auth]
      summary: Revoke a refresh token
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties: { refresh_token: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/OkEmpty' }
        '401': { $ref: '#/components/responses/Error' }

  # ── cases (3002) ───────────────────────────────────────────────────────────
...
  /billing/subscribe:
    post:
      tags: [billing]
      summary: Subscribe a lab to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lab_id, plan_id]
              properties:
                lab_id: { type: string, format: uuid }
                plan_id: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Subscription' } }
  /billing/usage/{lab_id}:
    parameters:
      - name: lab_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [billing]
      summary: Get a lab's current usage
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/UsageInfo' } }
  /billing/check-quota:
    post:
      tags: [billing]
      summary: Check whether a lab is within quota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lab_id]
              properties:
                lab_id: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/QuotaCheck' } }

# ───────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    OkEmpty:
      description: Success envelope with null data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
  schemas:
    Envelope:
      type: object
      description: Standard response wrapper used by every endpoint.
      required: [status]
      properties:
        status: { type: string, enum: [ok, error], example: ok }
        data: { description: Payload (shape varies by endpoint), nullable: true }
        error: { type: string, nullable: true, example: null }
    Role:
      type: string
      enum: [doctor, lab, admin]
...
    RegisterRequest:
      type: object
      required: [email, password, role, name]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        role: { $ref: '#/components/schemas/Role' }
        name: { type: string }
        organization: { type: string, nullable: true }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    AuthTokens:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string, format: uuid }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, example: 3600 }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        name: { type: string }
        organization: { type: string, nullable: true }
...
    Subscription:
      type: object
      properties:
        id: { type: string, format: uuid }
        lab_id: { type: string, format: uuid }
        plan_id: { type: string, format: uuid }
        status: { type: string, example: active }
        started_at: { type: string, format: date-time }
    UsageInfo:
      type: object
      properties:
        lab_id: { type: string, format: uuid }
        month: { type: string, format: date, example: "2026-05-01" }
        plan_name: { type: string, example: bronze }
        monthly_limit: { type: integer }
        current_usage: { type: integer }
        remaining: { type: integer }
    QuotaCheck:
      type: object
      properties:
        allowed: { type: boolean }
        remaining: { type: integer }
        monthly_limit: { type: integer }
        current_usage: { type: integer }